requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
django-cors-headers==3.10.0
django-filter==21.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.2
phonenumbers==8.12.36
Pillow==8.4.0
pluggy==1.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
